refactor(phase-4): split uptime extension into extensions/uptime/ + features/uptime/
Break the 1026-line extensions/uptime.py into mixin-based modules and extract
MongoDB I/O into features/uptime/ matching the Phase 4 pattern.

- features/uptime/models.py: MonitorConfig dataclass, STATUS_MAP, normalize_status
- features/uptime/repository.py: UptimeRepository (per-guild MongoDB persistence)
- extensions/uptime/_common.py: shared config and credential helper
- extensions/uptime/socketio_client.py: SocketIOMixin (connection, event handlers, dispatch)
- extensions/uptime/monitors.py: MonitorsMixin (monitor discovery + autocomplete)
- extensions/uptime/notifications.py: NotificationsMixin (state-change embeds)
- extensions/uptime/commands.py: CommandsMixin (/uptime setup/remove/list)
- extensions/uptime/tasks.py: TasksMixin (push heartbeat + REST fallback poll)
- extensions/uptime/__init__.py: UptimeExtension glue (composes mixins)
